Another depth fix and added an alpha multiplier
Fixed the viewport having a min and max depth of 0 causing depth rendering issues. Updated the rasteriser states to be created with GetAddressOf instead of using &. Added an alpha multiplier to the Appearance class and the constant buffer allowing for the transparency of an object to changed. Added a blender state to visually see the transparency effect. Added saturate the pixel shader's return value to guarantee that its values are between 0 and 1.