Updated texture transformation
Instead of giving a multiplier to get the desired frame width or height the SetTexCoords function now takes the number of frames for x and y. It uses them to calculate the multiplier instead of doing it manually. The last two arguments now tell it what frame in x and y to use. Also added a alpha multiplier value to the CreateGameObject function. Created a new folder for the texture files and updated Test2.dds.