Added the ability to add and remove game objects
Right clicking with the mouse now creates a new default game object with the Test.dds texture. While moving the object if the middle mouse button is clicked it now deletes the object. Made the code for this all in one if else if statement instead of having multiple if statements.