Updated mouse picking and snapping
Mouse picking is now more accurate as the mouse now uses a sphere instead of a box. Created the SnapCoordinatesToGrid in MousePicking which converts an X and a Y value to the nearest coordinates on a grid. Also made it so an object cannot be created on top of another object.