Got writing to and loading from a json file working
Added a write function to physics so it is now writes its information to a json file. The files can now also contain density and friction values. Fixed the write function in the appearance class not writing the alpha multiplier values of objects. Made ZoeLevel.dds the default background texture.