Can now change object textures through the UI
Added a text box in the appearance UI section where an alternative texture can be set. Made it so if an invalid texture file path is given this way it doesn't crash. Also made it so the first object's texture (should be the background object) is saved as the background texture in the json file.